Refactor clientDoc handler from proxy logic
